Branch for computing Hodge gradings
The Hodge filtration of a (g,K) module is canonical. We want to
compute the associated graded, written in terms of K-types. A formula

\sum a_i(q)mu_i

where each mu_i is a K-type indicates that mu_i has multiplicity
a_i(1), and its multiplicity in degree k of the Hodge grading is the
coefficient of q^k in a_i(q).

We define a type qParamPol to handle this. See qParamPol.at.