Working on implementing q_deformation_terms in qdeform.at
This requires computing the last column of the inverse of an upper
unitriangular polynomial matrix. This is done in deformation_terms.at
but only for a mat (the KL matrix evaluated at q=-1).

This has been extracted into the function
last_column_mat_polynomial_inverse.

compress (in deform.at) needs to be modified to return an i_poly_mat
(not just a mat).