Add a version of sort with general predicate, and do some small repair work
The general version of sort still sorts integers, but the supplied predicate may
interpret them as indices into any row, sinding a sorting permutation for it.
In passing the existing sort functions were slightly rewritten, and |merge_u|
repaired (its use of a sentinel value was not appropriate for unique merging).