Fixed bug in qdeform.at
This was due to the bug in deform which was fixed in commit c7de0b9d3.

q_recursive_deform appears to be working at least in SL(n,C) (n<=4)
and Sp(4,C), in the sense that it gives positive coefficients.
See the command qtest@(RealForm,int) in qdeform.at.