Fix bad typo in sl_list.h: rvalue designator was in the wrong function header
The methods |simple_list::splice(const_iterator,list&)| and its version with
|list&&| should run the same code, which should be done by having the latter
call the former. However, the |&| and |&&| in the headers were switched, making
the former call itself recursively; not a great idea. It happened that these
methods were never called in the Atlas software, but better to fix it anyway!