Make sl_list.h do the right thing for copy and move assignment with allocators
We finally figured out how the |propagate_on_container_copy_assignment|
allocator trait and its "move" relative are to be used, and implemented that for
|simple_list| and for |sl_list|. In conclusion things work best if propagation
follows ownership, that is is on copy assignement (which does not atempt to
transfer ownership) the allocator is not propagated (so nodes can be recuylced),
but on move assignment it is propagated (so that nodes can be moved). But the
code needs to respect what the allocator trait says, and to rebuild from scratch
when nodes cannot be recycled respectively moved.

We also textually moved allocator accessing methods upwards, for easier reading.