Improve |sl_list::insert| into |sl_list|, and remove the |ensure| helper class
Insertion is fastest at the end of a list (the null pointer of the newly
constructed node can remain) and this can be explioted by (1) taking aside the
case where insertion is at the end, and (2) in the remaiing case insert into an
empty list first, then splice in this list at the right position. The latter
case now also ensures roll-back when memory runs out during the insertion.