Modify |type_perm| by sorting sub-ranges; requires subtle change to range data
When sorting a sub-range of a list, the iterator marking the end of the range
needs to be adapted. The sorting procedure itself takes care of the ending
iterator passed to it (by reference), but cannot update other copies of the same
iterator, notably if serving as starting iterator of another range. A small but
tricky ad hoc change to the way we represent sub-ranges, namely by complementing
a start iterator by its dereferenced valuem allows us to solve the problem here.

We left the solution absed on overwriting node contents after sorting instead of
rearranging the links in place (but oncompiled) for reference.