Implement 3-cycling of links in sl_list.h systematically by two swaps
It probably does not make much of a difference, but two swaps of unique pointers
is both more compact to write and more likely to produce optimal machine code
than a sequence of four move assignements (probably an optimiser will see that
that the latter, out old solution, can never invoke any destructor, but for the
two swap method there is no question of destructin in the first place).

In passing make some tweaks for readability, which should not have any effect.