Replace used of mirrored lists by types derived from |std::stack| instances
Logically the cases in question are stacks, but we needed some operations that
are hidden by the |std::stack| interface, so used the mirrored list type that
would have been used by |containers::stack|, had it been used. This decreases
readability; the proper solution is to define a type derived from a |std::stack|
instance, and add some methods that bypass the hiding it establishes, to allow
the extended functionality but not everything else the underlying container has.