Clean up the way types are specified in sl_list.h, using embedded types more
We try to specify types by referring to the place that determines the right one,
often from a |..._traits| structure, rather than spelling them out explicitly.

This has the potential to vex the reader who tries to figure out what it all
boils down to, but in principle makes for better interoperability with for
instance allocators that work with unusual |pointer| types.