Move definition of |stack| and |queue| into sl_list.h, allow initialiser lists
The actual definition if the |stack| and |queue| classes should not be given in
the forward-declaration file sl_list_fwd.h, nor in Atlas.h, so it is moved to
sl_list.h, but without the "incompletecpp11" parts. We do deviate from the |std|
container adapters by allowing a queue constructor with initialiser list, useful
(and used) in particular to create queues primed with an initial element.