Have |containers::queue| and |stack| impose their choice of list container type
Previously they installed a new default to override the default from their |std|
counterparts. But makes little sense to use our versions for anyone who is going
to override the new default, so we might as well make the list container type a
fixed choice for our versions of the abstact type (which are no longer adaptors).