Working on computing component groups of centralizers of nilpotent orbits
This is mostly done on the annegret branch. The main thing which is
different here is using maybe_conjugate and is_conjugate from
conjugate.at in place of running over all of W. This occurs in the
functions pseudo_Levis and Kac_diagrams_with_centralizer.

Also modified print_component_info to have nicer output.

Note: print_component_info@ComplexNilpotent now uses tabulate from
tabulate.at, which has been updated (a small change) on this
branch. So for this to run on the annegret branch you'll need this
version of tabulate.at