Adding significant functionality to character_tables.at
These are the character tables for classical groups. Want to be able
to compute directly with them, rather than using W_characters.at,
which has more functions but *not* the complete character table. Added
some fields to the character_table type.

Also re-organization of files, for example renamed
strip_to_partition@[int] in combintatoric.at to delete_trailing_zeros
and moved it to misc.at