Add a simplified version |block_minimal| of |param_block|, intended to replace it
These blocks no longer store the infnitiesimal character, which eould need to be
stored separately in those places where it is desired. The intention is to have
just the information necessary for the abstract structure of the block, so that
Kazhdan-Lusztig tables can be shared across isomorphic blocks.

In order to achieve this independence of |gamma|, the |y| values are stored
using a representative of $\lambda-\gamma$ (a |RatWeight|) rather than of
$\lambda-\rho$ (a |Weight|); $\lambda$ itself is not something defined for the
|block_minimal| elements. The twist action, which will be necessary to define
extended blocks, is defined by a linear action on $\lambda-\gamma$.