Major changes related to unipotent calculations, especially character tables and springer tables of exceptional Weyl groups
The character tables for E6 and E7 are working.
The conjugacy classes are sorted lexicographically by
(order,class size,det,trace) and also using the power map.
The tables themelves have been computed using magma.

This includes the springer table for E6, but not E7,
so we can (and have) computed weak packets for E6.