Fix a bug in axis.w (loop type checking), calling |specialise| in wrong direction
This could lead to assertion failure when a loop expects a tuple component type
while the expression looped over has undertermined array type '[*]'. That
scenario no longer not breaks atlas (most likely giving an error while checking
the loop body instead) though the solution is still unsatisfactory, as the
information that an array of tuple type is required is not exported.