Introduce method |type_expr::is_unstable|, to find partially undetermined types
Such types are type patterns with some '*' remaining, which when interpreted as
type of a subexpression may be undesirable. Notably potential future
specialisation of such a type could invalidate overloading decisions, and in
cases where the type is copied from that of a variable, a specialisation of the
copy would not be recorded in the type of the variable, possibly giving
inconsistent behaviour of the type checker; in such cases the occurrence of a
type |t| satisfying |t.is_unstable()| should lead to an error stop.

In passing, type equality and inequality test are made into methods rather than
external functions; their implementation is best expressed using private fields.