Major rewording in axis source documentation, some error message improvement
A lot of source clarification; in the process, some sloppiness in error
messaging (for misplaced break and return) was repaired. The only real code
change involves suppressing '*' from ever being reported in a balance error, as
for instance for the illegal nested list '[[0,true],[false,1]]' (repeating the
same type has not been repaired though; it would require more code changes).