Fix long standing bug in extended KL computations: endgame case forgot link signs
The code in |ext_kl::KL_table::do_new_recursion| computes a sum that equals a
linear combination of the desired $P_{x,y}$ and some $P_{x^s,y}$ where $x^s$
represents 0, 1 or 2 ascent(s) of $x$ by $s$. The latter term(s) is/are known,
since they involve values $x'$ closer to $y$ than $x$. In most cases they are
taken into account by finding $P_{x^s,y}$ in |cy[sx]|, and multiplying it by its
coefficient |T_coef(s,x,tx)|, the latter integrating the sign on the link from
$x$ to $x^s$. However in the case called "endgame" (single-valued imaginary
ascent) there secretly also is a contribution of the sibling |s_cross_x| of |x|,
which not being an ascent of |x| does not have its $P_{scx,y}$ among the
values available in |cy|; however some other generator $t$ allows an easy-case
expression for $P_{scx,y}$ in terms of (1 or 2) values that can be found in |cy|.
That expression is computed on the fly (not using the standard mechanism for
doing it, as that would not look in the local array |cy| for previous values),
but in doing so, the signs on the links involved were not taken into account.

The proper fix was to tack multiplication by |epsilon(t,s_cross_x,x_prime)| onto
the polynomial |cy[x_prime]| (for various |x_prime|) in three places.

The bug caused erroneous results in very rare cases, which explains why it
passed under the radar for so long; eventually E7 computations revealed it.