Fix minor but very old bug in identifying real forms in unequal rank D4
The code did not take into account that one could be confronted with a D4
equipped with an unusual diagram involution (and hence position of the imaginary
roots among the simple ones), which sometimes led to the quasisplit form being
mislabelled so(7,1) (or worse?). Since other simple types can only have the
standard involutions if any, this was solved by an explict D4 exception.