Improve performance of extended block construction in trivial cases
When the twist does not affect any nodes of the Dynkin diagram and all the
elements of block are twist-fixed, then presumably the extended block is really
isomorphic to an ordinary block, so we skip the |check| call setting link signs.

In an example case this speeds up the unitary computation by a factor 3, though
it still remains significantly slower than for a similar equal-rank parameter.