Extensive changes to the format of the weak_packet algorithms.
The return value of weak_packet@(...) is now

([([Param],[WCell])],[(int,int,Param)]):

The first argument is a list of pairs ([Param],[WCell]), where [Param]
is a list of parameters in a block, and each [WCell] is a cell for
that block.

The second is a list of unipotent parameters, where (b,c,p) refers to
a unipotent parameter in block #b, cell #c.

To extract just the parameters from this do:

set wp=weak_packets(G) (or other variants)
set unipotents=parameters(wp)

and

export(unipotents)

to write a list of atlas-readable parameters.

There were some changes to Vogan-dual.at and nilpotent_orbits.at