Improve internal deformation algorithm functions in some relatively small ways
The main loop over |survivors| in |Rep_table::deformation_terms| had some strange
imperfection: it forgot to use the |remap| array as was intended, it recomputed
an index |i| that was just equal to the current loop index |pos|, and it
computed contributions even if the coefficient |c_cur| by which everything will
be multiplied should happen to be zero.

All this has been improved, and attention is given to not repeatedly calling
|block.sr|, but rather storing the |StandardRepr| values found the first time
for later reuse, and their hash values. In the end, the |survivors| list is no
longer looped over, replacing it with a counted loop over indices into it,
working with element hash values, and at the end their  |StandardRepr| values,
from this temporary storage. Also do some restructuring and comment improvement.