Remove needless complication in |Rep_table::twisted_deformation|
For unclear reasons, the first block needed in the deformation was computed
separately outside the main deformation loop, though effort was done to not
recompute it inside. Maybe we wanted this block (and its twisted KLV
polynomials) to survive longer, but this is silly as it is local to the function
anyway, so survival is relative. This change prepares more important speed-up.