Make |Rep_table| methods |deformation| and |twisted_deformation| more efficient
The main point is to avoid generating a block for parameters whose information
is already stored in the tables. This situation may arise inside a single call
of the (twisted) deformation function, even though the deformation terms at
different reducibility points belong to different blocks, and the deformation
terms at a single reducibility point can all share a single block construction:
due to recursion, those deformation terms again get deformed, and during that
process many may end up generating a same block independently. Now if a
parameter is encountered that already appears in the tables, no block is
constructed at all, but a special (new) version of |deformation_term| (or its
twisted companion) is called that directly operates parameters from the stored
polynomials, all of which are also already known in the tables.