Add a class |StandardReprMod| with hash behaviour taking |gamma| modulo $X^*$
The functions for equality and hashing have been changed to ignore translation
of the infinitesimal character by integer vectors (elements of $X^*$).

This may be used to store and look up information for a whole translation class
of parameters. Since a representative with full infinitesimal character is still
recorded, it would be possible to apply a shift to it after lookup, as required.