Add a function |deformation_terms| computing from a |block_minimal and |gamma|
The idea is to avoid the existing |Rep_table::add_block| methods, whose repeated
use will fill up the table with many possibly similar blocks and the polynomials
associated to their elements, and compute everything from a |block_minimal| that
could be common to a lot of blocks. The main issue treated is dealing with
condensation of the data to a set of final elements that may differ among blocks
sharing a |block_minimal| due to different singular subsystems. The storage and
lookup of a common |block_minimal| in a |Rep_table| still needs implemenation.

For the moment it looks like this may be trading memory for quite a bit of
repetitive computation (each call to |deformation_terms| has complexity the size
squared of the |block_minimal| interval up to the requested element), but once
functional, statistics and profiling may show ways to trade back some time
without exploding memory usage again.