Repair and change interface (|sl_list::emplace_back|) of singly linked lists
The repair is an mistyped identifier in the apparently never yet used methods
|simple_list::unique(BinaryPredicate)| and |sl_list::unique(BinaryPredicate)|.

The interface change is to have |sl_list::emplace_back| and |sl_list::push_back|
return a reference (not an iterator) to the created object. The motivation is
that without such a return value, getting such a reference would require holding
a copy of the |end| iterator from before the call and dereference it afterwards;
when used in the initialisation of a (reference) variable, that would force the
temporary iterator have the same lifetime as the variable initialised, not good.