Provide |ext_block| with a |unique_ptr| member to hold its own |ext_kl::KL_table|
This makes extended blocks similar to ordinary blocks that also store their own
table of KL polynimials if one has been generated. To facilitate this change,
the class |ext_kl::KL_table| is made more self-contained by having the
possibility to own its array of polynomials, which so far had to be defined
separately and passed to the constructor. This argument is now changed from a
reference to a pointer, so that it can be given a null pointer instead,
idicating that the constructor should create its own owned array.