Modify |block_minimal| to ensure that |gamma_lambda(z)| is $-\theta$ fixed
The method |block_minimal::gamma_lambda| did not always return a $-\theta$ fixed
rational vector, which is a consequence of |block_minimal| stocking at the end
of its construction the |TorusElement| $y$-values as-is rather than converting
them to bit vectors |y_bits| as in the |param_block| construction. This is now
changed by storing for each $y$ value a lifted |RatWeight| that is explicity
projected onto the $-\theta$ fixed subspace. The fact that the projection is in
the required $X^*$ coset (that of $\gamma-\rhos$) may not be evident (it is a
property that modifications to the |TorusElement| during the block construction
must respect), but it is already being tested in |param_block::compute_y_bits|
that is in use for years, so it is rather certain that it holds in practice.