Correct a sign in |star| function of block_minimal.cpp; had forgotten to flip it
In the passage from the use of a member |lambda_rho| in |ext_block::param| to a
member |gamma_lambda| in |ext_block::paramin|, the contribution of |lambda| is
in the opposite sense, so that changes to this component shoudl change sign.
This change had been forgotten in a single term, which caused trouble.