In function |repr::contributions| combine rather than concatenate lists
Combining takes place at double-values descents, if there any like terms in the
lists, they are combined and their multiplicities added; this is basically
sparse polynomial addition. Maybe in highly singular blocks this prevents
exponential increase (in terms of the maximal length in the block) of the length
of the lists that would occur if the lists are simply concatenated, though it
seems unlikely this will be a major limiting factor.