Fix a wrong |assert| in block_minimal.cpp, and improve some nearby code
The error was that a test for not being in the image lattice of |theta-1| that
was applied to just the numerator of a rational vector without normalisation;
the parasitic factor from the denominator made it wrongly be in the image.

The nearby code had a curiously too early introduced local variable, leaving it
undefined where it was not needed anyway. This was fixed with no pain, as was
the code in ext_block.cpp from which this had been originally copied.