Change |sum@(RealForm,[ParamPol])| so memory leak does not bite statistics code
Due to a known memory leak for recursive user functions, a recursive function
that in its environment catches some value provides an eternal reference to that
value, making it indestructable. If that value is or is dependent on a real
form, that storage for that real form is never released. This was in particular
the case for |sum@(RealForm,[ParamPol])|, which used a locally defined recursive
function that has the arguments of |sum| in scope, even though never uses them.
This function is used by |is_unitary| in the case of an unequal rank group.

The problem has been patched up by rewriting |sum@(RealForm,[ParamPol])| in such
a way that the recursion is made before the real form or any |ParamPol| for it
are seen, so that no sticky reference to them will prevent the real from from
ever being cleaned up. Changing the interpreter in such a way that recursion
does not necessariy give such a memory leak would be a better solution though.