Change |sum@(RealForm,[ParamPol])| so memory leak does not bite statistics code
The statistics diagnostics are printed by the |Rep_table| desctructor, but due
to a known memory leak for recursive user functions and the specifics of the
library code defining |is_unitary|, a call to that function for an unequal rank
group made the |Rep_table| for that group immortal, disabling diagnostics for
that group (even when trying to create a fresh copy of the same group).

The problem has been patched up by rewriting |sum@(RealForm,[ParamPol])| in such
a way that the recursion is made before the real form or any |ParamPol| for it
are seen, so that no sticky reference to them will prevent the real from from
ever being cleaned up. Changing the interpreter in such a way that recursion
does not necessariy give such a memory leak would be a better solution though.