Fix problems that appear to have existed on 'master' after merge with 'latest'
There were a number of cases of missing includes, or reference to a function
under a name used in misc.at while a different name was used in the files
actually included (|strip_to_partition|, and |semisimple_rank|).

There remains a sorry case of obviously broken code (flagged by comments) in
W_characters.at, but which could not be removed because other library files
depend on it. Some parts of basic.at affected by this commit are also dubious
(questions of how to define "simple" root data, and "highest_root" variants).