Define the |common_block| constructor for the partial case
This is largely inspired by the |param_block| constructor for the partial case,
but a bit complicated by the need to produce |y_entry| values for |y_hash|,
while it would be easy to everything in terms of |y_bits| that are readyly
available in our current approach. Once things work correctly, code should be
rewritten so as to do without |y_hash| altogether (espcially in |lookup|).