Remove code that shortcuts partial block recursion when hitting an existing block
This code is a lot of effort that moreover necessitates elaborate storage to
ensure no unnecessary recomputation is done. It will be more convenient to just
ignore the existing block and rely on the storage speeding up the recursion in
the case that no elements of the block are previously recorded.