Make exended blocks and twisted KL computations even more partial block tolerant
Very painfully try to do as much as possible even when close to the edge of a
partial block. The direct reason for doing this was that if a link is tuned from
one end, one should not refrain from doing the same from the other end under the
pretense that some other link is over the edge. This means we cannot simply
abandon our work just because we encounter the |UndefBlock| virus somewhere.