Clean up lattice.at, remove |intersection_plus| which was badly specified
The extra return value was described to be a chopped off part of the kernel of
A##B, but it was in fact so obtained only after applying |image_lattice| to each
of the matrices. This operation changes the kernel of A##B, and may even change
the number of columns of A and/or of B, and thus the size of kernel generators;
in the end the extra value is quite hard to relate to the original matrices.