Rewrite the common_block full block constructor, avoiding use of |nblock_help|
While this follows the general (very long) sequence of operations that used to
be done in the full block constructor, some attempt has been done to streamline
the process and work as much as possible with |StandardReprMod| values, not
converting them to and from |(x,y)| pairs in the block all the time.
Nonetheless, after a Cayley descent, we apply imaginary-cross orbit generation
to get a complete fiber of |x| values and then |repr::StandardReprMod::build|
the block elements form the appropriate set of |(x,y)| pairs; this is necessary
because one would not get the imaginary compact |x| values from Cayley descents.