Avoid linear search |Rep_table::block_list| by having |place| store iterators.
This is a rather subtle optimisation, based on the observation that each block
element, through the |Rep_table::mod_hash| tables, gives access to a |place|
entry. If instead of a pointer to the block it stores a list iterator giving
access to the block, then we can use that to |erase| the block after being
swallowed into a larger block without having to go through |block_list|. This
does imply worry about the validity of iterators when a node is erased; we
achieve this by adapting in |Rep_table::block_erase| the |place| fields of any
elements in the following node that might refer to the iterator to them that is
being invalidated by the erasure. To get this to work safely, we construct the
new block in a separate singleton list, splice it in place after all erasures
have taken place, and finally set the iterators in |place| for its elements.