Rename |KLContext| to more appropriate |KL_table| througout.
This is more consistent with terminology used elsewhere, where "context" usually
means a lightweight object that groups a number of values fixed over particular
kind of computation (like that of a block), so that these computations can access
them wihtout having to pass them around to all need places explicitly.

On the other hand, objects that retain lots of previously computed values are
better called tables, like |Rep_table|, and |KL_table| is of this nature.