Replace |KL_table::fill_limit| by a bitset |d_holes| of positions needed filling
This prepares the situation where adding block elements to a block in which part
of the KL polynomials have been computed leaves a number of holes in the table,
without invalidating the results already known. The convention is that anything
above |d_holes.capacity()| has not been touched Therefore so |d_holes.front()|
replaces the old value |fill_limit|. A successful computation of an entire
column is commited by clearing the corresponding but in |d_holes|; this we may
actually retain some computation in case an error occurs halfway.