Do recursions in KL computations using lists of |Mu_pairs|s rather than vectors
This is a bit more flexible, since the lengths of these lists are not known in
advance, and appears to not be slower (no linear searching is needed). Moreover
by keeping the block element $x$ and $\mu(x,y)$ together from the start, we
avoid a binary search for $x$ for the polynomial holding that coefficicent.

Cases of insertion-sorting part of the |MuColumn|, namely the part coming from
the |down_set| of |y|, are replace by a list |merge| call. (The sorting or
merging is necessary because the |down_set| elements, though of maximal length
among the |mu| contributors, need not be larger than all others.)