Make |deform| and |full_deform| ensure their parameters are dominant first
Before the common blocks came to be used, |deform| called |normalise| on its
argument, and not doing so any more in the common blocks code is confusing and
dangerous, as it even produced polynomials with non-standard terms down the
line. We fixed it by calling |normalise| in |Rep_table::lookup| and
|make_dominant| in |Rep_table::lookup_full_block|. For consistency we make
|full_deform| robust against non-dominant arguments by calling |normalise| (as
it does not use |lookup|). The twisted variants are exempt, as normalisation
might involve a sign flip of the default extension, so those cases need to be
handle with more explicit care; their wrappers require "final" parameters.

A consequence of this change is that |Rep_table::lookup| and |lookup_full_block|
now take their |StandardRepr| argument by modifiable reference, so that changes
to the infinitesimal character can be picked up by the caller (which is
essential for reconstructing the looked-up parameter in the block found). Several
wrapper functions now |get_own| their |Param| argument to then allow correctly
using |lookup| or |lookup_full_block| (without risk of corrupting a copy).

We also improved basic.at a bit, and included a test script in the repository.