Fix use of |expr| (within |assert|) after explicit destructor call.
This error was revealed when compiling with both debugging and optimising on,
since the optimiser simply removed the instruction whose purpose was to leave the
|expr| object in a well determined state, since inspecting a destructed object
is undefined behaviour anyway. This was solve by doing the same work in a new
non-destructor method, called by the destructor but also callable independently.