Ensure that |common_block| uses consistent ordering, so embedding is monotonic
We want the |swallow| method for |common_block|s to preserve stored |KL_table|
information, and in order to be able to do that without having to re-sort the
|d_KL| and |d_mu| vectors afterwards (in order to ensure that binary searching
applied to them continues to work), we shall ensure that the renumbering
|embed|, passed to the |swallow| methods and describing how old positions in a
sub-block map to new positions, is always increasing.

We achieve this monotonicity by having |common_block| constructors ensure a
fixed ordering of the |info| elements: by |length| first, then by |x|, and
finally by the |y_stripped| characterising attribute of |y| values. This
requires some special effort, as the |y_stripped| attribute is not available in
the |Block_base| class base, so the constructors make sure that for a given |x|
(therefore fixed involution) the |y| fields reflect increase of |y_stripped|
attributes. In the partial block constructor this can be achieved by ordering
the |y| values for a given |x| before matching them in |y_hash|, while in the
full block constructor this is achieved by renumbering |y| fields after the fact.