Add |pop_splice_to| methods to |containers::queue| to move front node to any list
This allows exploiting the fact that the |containers::queue| uses list nodes to
store its items: instead of copying the front node and then popping it, one can
now splice it into another |containers::sl_list| or |containers::simple_list|
without any copying. For convenience a reference to the moved item is returned.