Make |KL_table| capable of incorporating the tables of one for a sub-block
This is realised by implementing a |swallow| method for |KL_table|, similar to
and called by the one in |common_block|. While |KL_table::swallow| has an actual
definition that is quite simple, its proper functioning depends on using
consistent orderings between different |common_block| constructions so that the
|embed| map passed to |swallow| will always be increasing. An |assert| checking
this property (when debugging is enabled) is included. In the first attempts to
implement |KL_table::swallow| such precautions had not been made, which led to
weird errors popping up in KL computations when sub-blocks had been swallowed.

A subtle point in the implementation here is the fusion of the tables of (old
and new) polynomials, which for efficiency requires a hash table, but |KL_table|
has no such member (instead such a table is created in the constructor as a
local variable). In order so be able to build such a hash table as local
variable of |Rep_table::add_block_below|, methods have been added in |KL_table|
and in |common_blocks| to allow it to bind to the |d_store| member of the KL
table. In doing so |KL_table| is giving up strict control of manipulation of its
data, since hash table operations modify their underlying storage vector; it is
understood that callers should only use the hash table to pass it again to the
|KL_table::swallow| so that it can integrate new polynomials more easily.