Make filling primitivization vectors on-demand, method now is |fill_prim_index|
The motivation is that with many small blocks of relatively high rank, one is
probably looking at a huge number of quite small elements in |d_prim_index|,
most of which are never consulted; leaving those empty could save a lot of space,
and also the time of filling them.

This change did provoke a use of |mutable|, as previously |const| methods now
can have a side effect of internal extension-on-demand. It seemed most readable
and least disruptive to keep these methods |const|, whence the |mutable|.