Move primitive-index generation in |KLSupport| to a separate |prepare_prim_index|
This used to be done on the fly by look-up methods, and the |d_prim_index|
member had to be |mutable| for this reason. Having |prepare_prim_index| called
explicitly is both cleaner (as it avoids having a |mutable| member) and more
efficient (since no on-the-fly-testing is needed). It does mean the method
|KL_table::swallow| should take care to call |prepare_prim_index| for any
descent for which some |d_KL| is transorted.

While doing this, we also augment |prim_index| table in |KLSupport| with a range
indication for its entries, the total number of primitives for the decent set.
This is not used currently, but will be useful in coming merge from 'latest'.