Changes making KL computation a bit faster again
Mainly this avoids working in a large, sparsely used (only at extremal elements)
array |klv| and instead restricting it to hold only the used elements, remapping
extremal block elements to the proper index in the array. This appears to speed
up computation by a surprisingly large factor (about a third).

Also the allocation of |klv| is now outside the loop over different KL columns
so that all use the same array without reallocation.