Sanitize and simplify error reporting, removing |KLError| and |NumericOverflow|
Underflow reporting in KL computations indicates a logic error, should it is
better served by an |assert| (which in the debugger will allow localisation)
than by throwing a |NumericUnderflow| error in the polynomial arithmetic which
is then relabelled |KLError| while adding $x$ and  $y$ coordinates as was done.

As a consequence sources files kl_error,{h,cpp} disappear, as does the type
|NumericOverflow| in error.{h,cpp}, and some code streamlining by removing catch
blocks was enabled. In passing we redefine |Degree| to be only |unsigned int|.