Remove misleading argument |delta| from method |common_block::extended_block|
The presence of this argument suggests that it could validly be different from
|inner_class().distinguished()|, but it cannot since the result is stored in a
manner that does not depend on the value of |delta| at all, and this is
meaningsless if the |delta| used should be different from what the sotrage
expects it to have been. For any |delta|, use |ext_block| constructor directly.