Make |ext_kl::swallow| actually do what it was meant to: import ext_KL values
Like many of these small changes, this is actually more subtle than it appears.
The most imporatant point here is to deal with situations where either the
sub-block being swallow or the new block it is embedded into have a private hash
table for storing actual polynomials or share an external one (only if both
share the same table can be ignore any adaptation/tranlastion question of
polynomila indices). Actually the current code forces the new block to have a
private table (since getting the shared table from the |Rep_table| to the point
of |ext:kl::KL_table| construction is not yet implemented), but the code in this
commit should handle that case, as well as other hypothetical cases, correctly.