Make shared polynomial hash tables by actually used in (extended) common blocks
This while common blocks can (and must if optionality of sharing is to be taken
seriously) provide their hash table through the |KL_hash| method, their extended
blocks store a pointer to the share extended polynomial table upon construction,
for the |swallow| method to use it later. The asymetry here between ordinary and
extended cases is somewhat disturbing, but can be explained notably by that fact
that ordinary blocks are thought to be often constructed indepentently of the
need to compute KL polynomials, so it seems better to wait passing the hash
table for storing polynomials until KL polynomials are actually asked for, while
extended blocks really are only demanded for their extended polynomials, so it
makes sense to store a pointer for the polynomial hash table in the extended
block itself, waiting there until it is needed for fillling the |KL_table|.

This change required some changes to method prototypes to ensure the possibility
of getting the necessary data to the point of use, ultimately |swallow| methods.