Fix printing of |Split| values to repect the high order bits of components
Already |Split| values do 32-bit signed arithmetic without overflow testing (on
each of the evaluations at $1$ and $-1$), which isn't good but hopefully
tolerable, but in addition reconstruction of the coefficients of $1$ and $s$,
which ends with a division by $2$, was done in only $32$ bits arithmetic, which
basically makes the high order stored bits disappear, so that distinct intenal
values could be printed equal. Doing the conversion in 64 bits avoids this bug.