Make |W_graph| return symetric graph data, as well as |W_cells| for each cell
The actual change is in the underlying internal funcion |kl::wGraph|, so that
the output from Fokko is affected as well. The computation of |W_cells| requires
an oriented (asymmetric) graph, which is obtained by pruning edge lists under a
$\tau$-invariant non-inclusion condition, and this transition is made internally
before calling the |graph::OrientedGraph::cells| method (Tarjan's algorithm) to
decopose the oritented graph into strong components.