Change action in W_reps to its transpose, and coherent_irreducible implementation
According to the description Jeff gave in filtrations.pdf, the action matrices
for the Hecke and coherent continuation actions differ only by the signs on the
diagonal, even though abstractly the generator actions are negated (the apparent
contradiction is explained by the fact that the Hecke action is expresed on a
length-parity flipped basis). The action in W_reps.at used to be the transpose
of the coherent continuation action, and is now redefine to not transpose, which
is facilitated by the symmetric representation of W-grphs that is now used. The
matrices in coherent_irreducible.at define (paradoxically) the Hecke action so
differ only in the diagaonal entries; the implementation which was cumbersome is
replaced by essentially copying from W_reps.at. The explicit computation of
length difference parity (and the arguments necessary for it) was removed, since
W-graphs are biparite by length parity of the vertices. Finally comments have
been added to clarify, and naming of functions has been changed to reflect the
precise roles of the functions.